Moved spaces to their own package
