+++ src/main/java/dk/dtu/compute/se/pisd/roborally/controller/FieldAction.java
@@ -21,7 +21,7 @@
  */
 package dk.dtu.compute.se.pisd.roborally.controller;
 
-import dk.dtu.compute.se.pisd.roborally.model.Space;
+import dk.dtu.compute.se.pisd.roborally.model.spaces.Space;
 
 /**
  * ...
