+++ src/main/java/dk/dtu/compute/se/pisd/roborally/model/spaces/CheckPoint.java
@@ -1,5 +1,8 @@
-package dk.dtu.compute.se.pisd.roborally.model;
+package dk.dtu.compute.se.pisd.roborally.model.spaces;
 
+
+import dk.dtu.compute.se.pisd.roborally.model.Board;
+import dk.dtu.compute.se.pisd.roborally.model.Player;
 
 public class CheckPoint extends Space {
     public final int Id;
