+++ src/main/java/dk/dtu/compute/se/pisd/roborally/model/spaces/ObstacleType.java
@@ -1,4 +1,4 @@
-package dk.dtu.compute.se.pisd.roborally.model;
+package dk.dtu.compute.se.pisd.roborally.model.spaces;
 
 /**
  * @author ZeeDiazz (Zaid)
