+++ src/main/java/dk/dtu/compute/se/pisd/roborally/view/ObstacleView.java
@@ -1,7 +1,7 @@
 package dk.dtu.compute.se.pisd.roborally.view;
 
 import dk.dtu.compute.se.pisd.designpatterns.observer.Subject;
-import dk.dtu.compute.se.pisd.roborally.model.Obstacle;
+import dk.dtu.compute.se.pisd.roborally.model.spaces.Obstacle;
 import javafx.scene.layout.StackPane;
 
 /**
